+++ app/src/main/java/br/iesb/songs/views/fragments/UserNameFragment.kt
@@ -1,7 +1,9 @@
 package br.iesb.songs.views.fragments
 
 import android.app.Activity
+import android.content.ContentValues.TAG
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -11,6 +13,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import br.iesb.songs.R
 import br.iesb.songs.view_model.LoginViewModel
+import br.iesb.songs.views.fragments.principal_fragments.LocationFragment
 import kotlinx.android.synthetic.main.fragment_user_name.*
 
 class UserNameFragment(private val type: String) : Fragment() {
@@ -28,6 +31,7 @@ class UserNameFragment(private val type: String) : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        Log.w(TAG, "TESTEGERE")
 
         if (type != "exist") {
             closeButton.visibility = View.GONE
@@ -57,6 +61,7 @@ class UserNameFragment(private val type: String) : Fragment() {
     }
 
     private fun dismiss() {
+        context?.let { LocationFragment(it).onResume() }
         val manager = activity?.supportFragmentManager
 
         manager?.findFragmentByTag("userName")?.let { it1 ->
